+++ STM_Shuttles/PLUDOS_Edge_Node/Core/Inc/stm32u5xx_it.h
@@ -55,6 +55,8 @@ void SVC_Handler(void);
 void DebugMon_Handler(void);
 void PendSV_Handler(void);
 void SysTick_Handler(void);
+void RTC_IRQHandler(void);
+void EXTI11_IRQHandler(void);
 void EXTI14_IRQHandler(void);
 void EXTI15_IRQHandler(void);
 /* USER CODE BEGIN EFP */